Refactor main into modules
